Refactor structure of split code
